Temporary file to understand docutils input processing.
In test_inline_markup, it was observed that the doctree passed to the writer
is different that the pseudoxml output in the reference docutils parser test.
A temporary sandbox file is being use to understand what part of the default
docutils machinery is responsible for altering the doctree before it gets
passed to the writer.