Fixes to paragraphs with odd start.
It was found that numerous aliases to other than `paragraph` body elements need
to be escaped. Some of that was fixed (and some of tests needed to be updated
with those escapes).

Fixed depper level line blocks and added tests for it.

Added support for an extra `title` directive.

Started support for the `math` directive.