Changed resolution of relative paths of `image` and `figure` directives.
Resolution of relative paths in (transitive) included contents is inconsitent in
docutils. The `include` directive resolves relative paths with respect to the
source RST file, while the `image` and `figure` directives keep the relative
path unchanged.

This commits extends the default `image` and `figure` directives by adding the
`source` attribute to the image doctree node and lets the RST writer use that
information (if available) to resovle relative path with respect to their
source RST file.