rst2html converter producing simple, human readable HTML output.
This converter builds on docutils' `html4css1` writer and ceases use of
stylesheet and template files.

The primary use of the converter is to generate HTML code for testing
the HTML parser. Having the path from rst to doctree to HTML to
doctree gives opportunity to test the two doctree representations for
equality/similarity.