Inception version of a YAML writer.
A YAML writer is to substitute docutils' pseudo-XML writer in getting a
string representation of a document tree. The plan is to use this as the
document tree representation in unit tests (of a HTML reader) to hold the
expected document tree. The YAML form can be easily converted to `dict`
and comparison run upon it.

YAML is preferred to JSON as it would be easier to write.