Inception version of a YAML parser.
The `YamlParser` is a counterpart to the `YamlWriter` and its purpose is
to recover a document tree from the YAML text output of the `YamlParser`.
This inception code is pretty simple and uses recursive calls to parse
through the YAML tree structure. Error conditions checking is minimal
and may lead to unexpected exceptions.