Unit test exercising indentation of a table under list item.
The test is presently failing because of wrong indentation applied to
the table nodes. Looking at the code, the existing concept of
indentation is somewhat akward as the `Writer.get_indent()` tries to
serve two purposes: 1) Return a line prefix to a downstream (child)
nodes in the document tree, and 2) return a line prefix to be used
for the text in a given node. The two prefixes would often be the same
but would differ on certain occassions (e.g. first line of a the first
paragraph node under a list item node).

Hence a potential fix should focus on separating the two use case of
the indentation prefix.