Proof of concept in fixing indentation.
Introduced a new `Writer.get_indents()` class method to supercede the
existing `Writer.get_indent()`. Proved the concept on the `table` node
and its failing `test_fixes`. The idea is to allow a list of indentation
prefixes such that the index in that list corresponds to the index of
a text line that will use the prefix. This way, first line can receive
a different prefix than the second and further text lines.

Future commits will try to extend the concept to other node types than just
the `table`. Specifically the `paragraph` node will ultimately exercise
the concept feasibility.